+++ storage/convert.go
@@ -1,7 +1,6 @@
 package storage
 
-
-func getBoundedLimit(limit uint) uint {
+func getBoundedLimit(limit uint64) uint64 {
 	if limit > maxLimit {
 		return maxLimit
 	} else {
